Fix: Ensure usePrayerContext is called within PrayerProvider scope
This commit addresses an error where `usePrayerContext` (called via the
`useAudioPlayer` hook) was invoked outside of a `PrayerContext.Provider`.

The fix involves the following changes:

1.  **New `PrayerContent` Component:**
    *   I created `src/app/prayer/_components/PrayerContent.tsx`.
    *   This component now encapsulates the main UI rendering logic, local UI
        state management (e.g., `searchTerm`, `isSidebarOpen`), and hook calls
        (including `useAudioPlayer`, `useWindowSize`, `useInterval`, and
        various memoized callbacks and effects) that were previously in
        `PrayerLyricsLoop`.
    *   `PrayerContent` consumes `PrayerContext` using `usePrayerContext()`,
        ensuring all hooks and child components within it correctly access
        the context.

2.  **Simplified `PrayerLyricsLoop` (`src/app/prayer/page.tsx`):**
    *   `PrayerLyricsLoop` is now a minimal component primarily responsible
        for initializing the `prayerReducer` (to get `state` and `dispatch`).
    *   It then renders `<PrayerContext.Provider value={{ state, dispatch }}>`
        with `<PrayerContent />` as its direct child.

This restructuring ensures that `useAudioPlayer` and any other part of your
application relying on `PrayerContext` are correctly nested within the
provider, resolving the "usePrayerContext must be used within a
PrayerProvider" error. The overall application logic and user-facing
functionality remain unchanged.
